\name{BuyFee}
\alias{BuyFee}
\docType{data}
\title{
Transaction Fee for Buying a stock
}
\description{
This represents the value in dollars in transaction fee for purchasing a stock.
}
\format{
  To change the ticker:
  BuyFee <- 1
}
\examples{
# To display current BuyFee
BuyFee
}
\keyword{datasets}
