\name{cashTicker}
\alias{cashTicker}
\docType{data}
\title{
Ticker Symbol for Cash
}
\description{
This is a ticker symbol used to represent cash in portfolio
}
\format{
  To change the ticker:
  cashTicker <- "CASH"
}
\examples{
# To display current value for cashTicker
str(cashTicker)
}
\keyword{datasets}
