\name{PortfoliosInfo}
\alias{PortfoliosInfo}
\docType{data}
\title{
Data frame contain information on portfolios
}
\description{
Data frame containing information on portfolios in simulator
}
\format{
  A data frame with the following 3 variables.
  \describe{
    \item{\code{PORTFOLIOID}}{Id of portfolio}
    \item{\code{ALGORITHM}}{name of portfolio}
    \item{\code{STARTAMOUNT}}{starting investment amount}
    \item{\code{PERFORMANCE}}{percentage return on investment}
  }
}

\examples{
## See TestSimulator.R
}
\keyword{datasets}
