\name{portfolioStocksChart}
\alias{portfolioStocksChart}

\title{
Plots Portfolio Stocks Chart
}
\description{
Compares performance of different stock prices in our portfolio
}
\usage{
portfolioStocksChart(data)
}

\arguments{
  \item{data}{
Portfolio Data
}
}
\references{
Wickham, Hadley (2009) ggplot2
Chambers, J. (2008) Software for Data Analysis
}

\examples{
# see TestVisualizer.R
}
