\name{PrintSimulatorState}
\alias{PrintSimulatorState}
\title{
Prints information on the current state of the simulator
}
\description{
This function displays information such as the time period, nummber of portfolios and other useful general information on the simulator state
}
\usage{
PrintSimulatorState()
}

\examples{
## See TestSimulator.R
}

