\name{readFromDB}
\alias{readFromDB}
\title{
Reads data from sqlite table
}
\description{
Reads data from sqlite table into a data frame
}
\usage{
readFromDB(dbFile, tablename, where = "")
}
\arguments{
  \item{dbFile}{
Database file path and name
}
  \item{tablename}{
Table name
}
  \item{where}{
SQL where clause
}
}

\examples{
## see TestVisualizer.R
}
