\name{schemaHoldingsTableName}
\alias{schemaHoldingsTableName}
\docType{data}
\title{
Database table name for Asset Holdings Info
}
\description{
Represents the table name for storing Asset Holdings Info
}
\format{
  To change this value: 
  schemaHoldingsTableName <- "holdings"
}

\examples{
## Setting this value
schemaHoldingsTableName <- "holdings"
}
\keyword{datasets}
