\name{schemaMarketTableName}
\alias{schemaMarketTableName}
\docType{data}
\title{
Database table name for Stock Market Daily Prices Info
}
\description{
Represents the table name for storing Stock Market Daily Prices Info
}
\format{
  To change this value:
  schemaMarketTableName <- "market"
}

\examples{
## Setting this value
schemaMarketTableName <- "market"
}
\keyword{datasets}
