\name{schemaPortfoliosTableName}
\alias{schemaPortfoliosTableName}
\docType{data}
\title{
Database table name for Portfolios Info
}
\description{
Represents the table name for storing Portfolios Info
}
\format{
  To change this value:
  schemaPortfoliosTableName <- "portfolios"
}
\examples{
## Setting this value
schemaPortfoliosTableName <- "portfolios"
}
\keyword{datasets}
