\name{schemaStocksTableName}
\alias{schemaStocksTableName}
\docType{data}
\title{
Database table name for Individual Stock Info
}
\description{
Represents the table name for storing Individual Stock Info
}
\format{
  To change this value:
  schemaStocksTableName <- "stocks"
}

\examples{
## Setting this value
schemaStocksTableName <- "stocks"
}
\keyword{datasets}
