\name{schemaTransactionsTableName}
\alias{schemaTransactionsTableName}
\docType{data}
\title{
Database table name for Transaction Info
}
\description{
Represents the table name for storing Transaction Info
}
\format{
  To change this value:
  schemaTransactionsTableName <- "transactions"
}

\examples{
## Setting this value
schemaTransactionsTableName <- "transactions"
}
\keyword{datasets}
