\name{SellFee}
\alias{SellFee}
\docType{data}
\title{
Transaction Fee for Selling a stock
}
\description{
This represents the value in dollars in transaction fee for selling a stock.
}
\format{
  To change the ticker:
  SellFee <- 1
}
\examples{
# To display current SellFee
SellFee
}
\keyword{datasets}
