\name{SimulateBroker}
\alias{SimulateBroker}
\title{
Simulates the execution of an order book on the stock market
}
\description{
This function simulates the execution of orders for a Portfolio on the stock market. It outputs dataframes containing updated asset holdings information and executed transactions.
}
\usage{
SimulateBroker(portfolioId, PortfoliosInfo, StockMarketDataset, OrderBook, HoldingsInfo, TransactionsInfo)
}
\arguments{
  \item{portfolioId}{
Id of Portfolio
}
  \item{PortfoliosInfo}{
Data frame containing information on portfolios in simulator
}
  \item{StockMarketDataset}{
Data frame containing stock market dataset
}
  \item{OrderBook}{
Data frame containing orders generated by trading algorithm
}
  \item{HoldingsInfo}{
Data frame containing information on asset holdings in simulator
}
  \item{TransactionsInfo}{
Data frame containing information on transactions in simulator
}
}

\value{
Returns a list containing two dataframes: HoldingsInfo and TransactionsInfo
}

\examples{
## See TestSimulator.R
}
