\name{TradeStrategyAnalyzer-package}
\alias{TradeStrategyAnalyzer-package}
\alias{TradeStrategyAnalyzer}
\docType{package}
\title{
Trade Strategy Analyzer
}
\description{
This package contains a trade simulator, a collection of trading strategies and visualization techniques. For a description of the package, please refer to TradeStrategyAnalyzer_WhitePaper.pdf included in inst/doc.
}
\details{
\tabular{ll}{
Package: \tab TradeStrategyAnalyzer\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-06-10\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Winnie Cheng <wwcheng@alum.mit.edu>,
Tirto Adji <tirto_adji@yahoo.com>
}
\references{
Please refer to TradeStrategyAnalyzer_WhitePaper.pdf included in the package.
}

\keyword{ package }

